Remember the identifier for coodinate when coordinate is out of the map boundry.
